+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Direct/indirect imports and listing imported packages
 - pin/unpin packages and unimport --all
+- rename import -> load, unimport -> unload
 
 ## [0.1.4] - 2025-11-24